segnet: Fix UNet loss and add a new network based on TernausNet
I've tried to match TernausNet architecture more or less but there
are differences in details that I don't care enough to correct for :)

The second model trains to a good result much more quickly. Ideally,
overfitting would need to be solved for it, but for now segmentation
seems to be in a good enough place to park.