segnet: multi box fml
This attempts to implement multiple boxes detected per cell using
anchors, IOU thresholds for ground truth losses etc.

This... doesn't work very well. There's probably bugs somewhere in the
code and this is just... complicated. This checkpoint is there so that
we don't lose the data, but this really does not work very well at all.